moved config into app
