+++ internals/bin/server.js
@@ -31,7 +31,7 @@ const WebpackIsomorphicTools = require('webpack-isomorphic-tools');
 const isomorphicConfig = require('../webpack/webpack-isomorphic-tools');
 global.webpackIsomorphicTools = new WebpackIsomorphicTools(isomorphicConfig)
   .server(rootDir, () => {
-    const config = require('../../src/config');
+    const config = require('config');
     const { port = 3000, host = 'localhost' } = config;
     require('../../src/server').listen(port, (err) => {
       if (err) {
